+++ CharacterDataEditor/Constants/VersionConstants.cs
@@ -1,8 +1,8 @@
-﻿namespace CharacterDataEditor.Constants
+namespace CharacterDataEditor.Constants
 {
     public class VersionConstants
     {
-        public const string CurrentVersion = Ver103;
+        public const string CurrentVersion = Ver110;
 
         public const string Original = "";
         public const string Ver094 = "Ver0.9.4";
@@ -12,5 +12,6 @@
         public const string Ver101 = "Ver1.0.1";
         public const string Ver102 = "Ver1.0.2";
         public const string Ver103 = "Ver1.0.3";
+        public const string Ver110 = "Ver1.1.0";
     }
 }