+++ CharacterDataEditor/Enums/PreviousVersions/MoveType.cs
@@ -1,0 +1,25 @@
+﻿namespace CharacterDataEditor.Enums.PreviousVersions
+{
+    public enum MoveType
+    {
+        BackwardThrow = 0,
+        ForwardThrow = 1,
+        Grab = 2,
+        SideSpecial = 3,
+        NeutralSpecial = 4,
+        StandingLight = 5,
+        StandingLight2 = 6,
+        StandingLight3 = 7,
+        StandingMedium = 8,
+        StandingHeavy = 9,
+        CrouchingLight = 10,
+        CrouchingMedium = 11,
+        CrouchingHeavy = 12,
+        JumpingLight = 13,
+        JumpingMedium = 14,
+        JumpingHeavy = 15,
+        UpSpecial = 16,
+        DownSpecial = 17,
+        CommandGrab = 18
+    }
+}